requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
sqlalchemy==2.0.41
beautifulsoup4==4.13.4
requests==2.32.4
langchain==0.3.26
openai==1.93.0
python-dotenv==1.1.1
faiss-cpu==1.11.0
sentence-transformers==4.1.0
pytest==8.4.1
pydantic==2.11.7
httpx==0.28.1
pandas==2.2.3
numpy==2.3.1
asteval
google-generativeai
langchain-google-genai
streamlit
langchain-community

